pyproject.toml
[build-system]
requires = ["setuptools>=69", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "retrollm"
version = "0.1.0"
description = "Standalone reproduction of a neural-guided retrosynthesis search pipeline with pluggable LLM meta-control"
readme = "README.md"
requires-python = ">=3.10,<3.13"
dependencies = [
  "numpy<2",
  "pandas>=1.5,<3",
  "tables>=3.8",
  "pyyaml>=6",
  "requests>=2.28",
  "tqdm>=4.60",
  "python-dotenv>=1.0.0",
  "onnxruntime>=1.16",
  "rdchiral>=1.1.0",
  "rich>=13",
]

[project.optional-dependencies]
dev = [
  "pytest>=7",
]

[project.scripts]
retrollm = "retrollm.cli:main"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]
